refactor: UserService의 메소드를  static으로 변경
