+++ src/main/java/service/UserService.java
@@ -8,17 +8,17 @@ import java.util.List;
 import java.util.Optional;
 
 public class UserService {
-    public void addUser(String userId, String password, String name, String email) {
+    public static void addUser(String userId, String password, String name, String email) {
         User user = new User(userId, password, name, email);
         DataBase.addUser(user);
     }
 
-    public boolean login(String userId, String password) {
+    public static boolean login(String userId, String password) {
         Optional<User> user = DataBase.findByUserId(userId);
         return user.map(value -> value.isCorrectPassword(password)).orElse(false);
     }
 
-    public List<User> showList() {
+    public static List<User> showUserList() {
         return new ArrayList<>(DataBase.findAll());
     }
 }
