+++ src/main/java/webserver/HandlerMapping.java
@@ -5,19 +5,19 @@ import webserver.handler.LoginHandler;
 import webserver.handler.UserHandler;
 import webserver.handler.UserListHandler;
 
-import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class HandlerMapping {
-    private static final Map<String, Handler> map = new HashMap<>();
+    private static final Map<String, Handler> HANDLERS = new ConcurrentHashMap<>();
 
     static {
-        map.put("/user/create", new UserHandler());
-        map.put("/user/list", new UserListHandler());
-        map.put("/user/login", new LoginHandler());
+        HANDLERS.put("/user/create", new UserHandler());
+        HANDLERS.put("/user/list", new UserListHandler());
+        HANDLERS.put("/user/login", new LoginHandler());
     }
 
     public Handler getHandler(String uri) {
-        return map.get(uri);
+        return HANDLERS.get(uri);
     }
 }
