+++ src/main/java/webserver/RequestParser.java
@@ -4,7 +4,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.FileIoUtils;
 import utils.IOUtils;
-import webserver.http.*;
+import webserver.http.HttpMethod;
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
+import webserver.http.HttpStatus;
+import webserver.http.cookie.HttpCookie;
+import webserver.http.cookie.HttpCookies;
+import webserver.http.header.HttpHeader;
+import webserver.http.header.HttpHeaders;
 
 import java.io.BufferedReader;
 import java.io.IOException;
