+++ src/main/java/webserver/http/cookie/HttpCookies.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.cookie;
 
 import java.util.Map;
